requirements.txt
easyocr
streamlit
numpy
Pillow
